refactor: simplify grammar and lexer rules
grammar and lexing rules were written on a tag/value pair basis
which was becoming a bit heavy to maintain and most likely
isn't as performant as it could be (however this hasn't been verified)

this refactor is an attempt to generalise standard grammar and lexer rules
while allowing exceptions where they are needed.

BREAKING CHANGE:

this refactor doesn't aliases RIS tags to "human-friendly" names anymore.
these aliases actually depend on the type of the reference, so these
aliases will become dynamic in the future.

for the time being each tag/value pair will become a key/value
pair of the reference object.

some of the values are also postprocessed to add some more
metadata to them. this isn't being documented in this commit