refactor: relaxation of grammar & fault-tolerant parser
The lexer has become difficult to manage.
The definition and priority of tokens were conflicting.

For example, I defined these two tokens:

- A "name" token is a sequence of <a-z+><,><a-z+> (e.g. Doe,John).
- A "value" token is a sequence of any character.

Because the "name" token is defined before the "value" token,
the parser would fail for a tag expecting a "value" token that contains
the same sequence as the "name" token. e.g.,

AB  - foo,bar
      ^
      failure: expected "value" token but saw a "name" token instead

In addition some vendors seem to produce RIS files that, to the
best of my knowledge, aren't compatible with the RIS specification
which admittedly is sometimes confusing.

For example I have seen this kind of RIS file:

TY  - JOUR
KW  - foo
ER  -

If your editor doesn't show you whitespaces,
the error is that the ER tag contains and extra space:

ER<SPACE><SPACE>-<SPACE><SPACE>

Whereas the RIS spec says that it should be:

ER<SPACE><SPACE>-<SPACE>

In this particular case it seemed unnecessary to have the parser fail.
It will definitely not be obvious to the user what the error is and it
so easy for the parser to just ignore that extra space.

I have therefore decided that the grammar and the lexer will
simply facilitate the parsing of RIS files and act less as validating agents.

BREAKING CHANGE:

The parser will not fail unless the RIS file doesn't follow the basic specification:

<TAG><SPACE><SPACE>-<SPACE><CONTENT>

For example the parser used to fail for this:

RP  - FOOBAR

Expected value for the "RP" tag is:

- IN FILE
- NOT IN FILE
- ON REQUEST (mm/dd/yyyy)

Now the parser will simply take the content as is and will make best-effort attempts
to make sense of the data.

In a nutshell do not use the parser as a validation tool anymore.