feat: release new parser
BREAKING CHANGE:

This changes the parser to support any tags
as long as they are correctly formatted.

All tags are put into arrays.

The ER tag is allowed to miss the trailing space
after the dash.

Special formatting rules for some tags are
supported but not enforced.

Other changes include

* There is no default export anymore.
  To parse content call the parser function.

* Log a message instead of throwing an exception
  when content cannot be parsed.